// certigon.employee.web/certigon.employee.web/src/app/models/employee.model.ts
export class Employee {
    constructor(init?: Partial<Employee>) {
      Object.assign(this, init);
    }
  
    id?: number;
    name = "";
    surname = "";
    city = "";
    department: Department  = Department.Development;
    isActive?: boolean;
  }
  
  export enum Department {
    Development = 0,
    Management = 1,
    HR = 2,
  }